requirements.txt
# FastAPI Backend Dependencies with OpenAI
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.24.1


python-multipart==0.0.6

# OpenAI Integration (REQUIRED)
openai==1.3.0

# Optional: For enhanced features
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
